Moved aux function to utils.
